Make shared data generators execute on-demand
Before this change, shared data generators (the ones which provide shared data such as deprecated WinAPI lists or the most recent Electron version) were executed before the analysis start, even if there were no executables which actually need these shared data for analysis. After the change, these data will be loaded on-demand, only of there is a rule and an executable which require the data for analysis.